Move middleware
