+++ app/Http/Controllers/DirtRally/ChampionshipSeasonEventController.php
@@ -12,8 +12,8 @@ class ChampionshipSeasonEventController extends Controller
     public function __construct()
     {
         $this->middleware('admin', ['except' => ['show']]);
-        $this->middleware('validateSeason', ['only' => ['create', 'store']]);
-        $this->middleware('validateEvent', ['only' => ['show', 'edit', 'update', 'destroy']]);
+        $this->middleware('dirt-rally.validateSeason', ['only' => ['create', 'store']]);
+        $this->middleware('dirt-rally.validateEvent', ['only' => ['show', 'edit', 'update', 'destroy']]);
     }
 
     /**
