+++ app/Http/Controllers/DirtRally/ChampionshipSeasonEventStageController.php
@@ -13,8 +13,8 @@ class ChampionshipSeasonEventStageController extends Controller
     public function __construct()
     {
         $this->middleware('admin', ['except' => ['show']]);
-        $this->middleware('validateEvent', ['only' => ['create', 'store']]);
-        $this->middleware('validateStage', ['only' => ['show', 'edit', 'update', 'destroy']]);
+        $this->middleware('dirt-rally.validateEvent', ['only' => ['create', 'store']]);
+        $this->middleware('dirt-rally.validateStage', ['only' => ['show', 'edit', 'update', 'destroy']]);
     }
 
     /**
