+++ app/Http/Middleware/DirtRally/ValidateEvent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\DirtRally;
 
 class ValidateEvent
 {
