+++ app/Http/Middleware/DirtRally/ValidateStage.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\DirtRally;
 
 class ValidateStage
 {
